Refactor sidebar into separate module
